+++ packages/base/src/features/story/presentation/storyPresentation.ts
@@ -1,0 +1,15 @@
+import type { IJGISStoryMap } from '@jupytergis/schema';
+
+import { STORY_TYPE } from '@/src/types';
+
+export function isVerticalScrollStory(
+  storyType: IJGISStoryMap['storyType'] | undefined,
+): boolean {
+  return storyType === STORY_TYPE.verticalScroll;
+}
+
+export function isColumnStory(
+  storyType: IJGISStoryMap['storyType'] | undefined,
+): boolean {
+  return !isVerticalScrollStory(storyType);
+}
