+++ packages/base/src/mainview/components/MainViewSpectaPanel.tsx
@@ -1,9 +1,17 @@
 import { IJGISLayer, IJupyterGISModel } from '@jupytergis/schema';
-import React, { RefObject } from 'react';
+import React, { RefObject, useEffect, useMemo, useRef } from 'react';
 
-import { StoryPresentationRoot } from '@/src/features/story/presentation/StoryPresentationRoot';
+import { useStoryMap } from '@/src/features/story/hooks/useStoryMap';
+import { getStoryPresentationMode } from '@/src/features/story/presentation/getStoryPresentationMode';
+import {
+  StoryPresentationDesktopChrome,
+  StoryPresentationMobileChrome,
+} from '@/src/features/story/presentation/StoryPresentationChrome';
 import type { IStoryViewerPanelHandle } from '@/src/features/story/StoryViewerPanel';
-import type { IListStorySegmentTransition } from '@/src/features/story/types/types';
+import type {
+  IOverrideLayerEntry,
+  IListStorySegmentTransition,
+} from '@/src/features/story/types/types';
 
 export interface IMainViewSpectaPanelProps {
   model: IJupyterGISModel;
@@ -32,21 +40,80 @@ export function MainViewSpectaPanel({
   onSegmentTransitionEnd,
   onSegmentTransitionChange,
 }: IMainViewSpectaPanelProps): JSX.Element | null {
+  const overrideLayerEntriesRef = useRef<IOverrideLayerEntry[]>([]);
+  const segmentContainerRef = useRef<HTMLDivElement>(null);
+  const {
+    storyData,
+    currentIndex,
+    setIndex,
+    handlePrev,
+    handleNext,
+    hasPrev,
+    hasNext,
+    activeSlide,
+    layerName,
+    showGradient,
+  } = useStoryMap({
+    model,
+    overrideLayerEntriesRef,
+    removeLayer,
+    addLayer,
+    isSpecta,
+  });
+
+  const presentationMode = useMemo(
+    () => getStoryPresentationMode(storyData?.storyType),
+    [storyData?.storyType],
+  );
+
+  useEffect(() => {
+    const el = segmentContainerRef.current;
+    if (!el || !onSegmentTransitionEnd) {
+      return;
+    }
+
+    const handleAnimationEnd = (e: AnimationEvent) => {
+      if (e.animationName === 'fadeIn') {
+        el.removeEventListener('animationend', handleAnimationEnd);
+        onSegmentTransitionEnd();
+      }
+    };
+    el.addEventListener('animationend', handleAnimationEnd);
+
+    return () => el.removeEventListener('animationend', handleAnimationEnd);
+  }, [currentIndex, onSegmentTransitionEnd]);
+
   if (!initialLayersReady) {
     return null;
   }
 
+  const chromeProps = {
+    model,
+    isSpecta,
+    presentationMode,
+    segmentContainerRef,
+    storyData,
+    currentIndex,
+    setIndex,
+    activeSlide,
+    layerName,
+    handlePrev,
+    handleNext,
+    hasPrev,
+    hasNext,
+    showGradient,
+    onSegmentTransitionChange,
+  };
+
+  if (isMobile) {
+    return <StoryPresentationMobileChrome {...chromeProps} />;
+  }
+
   return (
-    <StoryPresentationRoot
-      model={model}
-      isSpecta={isSpecta}
-      isMobile={isMobile}
-      onSegmentTransitionEnd={onSegmentTransitionEnd}
+    <StoryPresentationDesktopChrome
+      {...chromeProps}
       containerRef={containerRef}
       storyViewerPanelRef={storyViewerPanelRef}
-      addLayer={addLayer}
-      removeLayer={removeLayer}
-      onSegmentTransitionChange={onSegmentTransitionChange}
     />
   );
 }
